+++ test/test_from_scratch.sh
@@ -4,11 +4,10 @@
 #Purpose: 
 #    Run set up and test of snppipline code on a specified test data set.
 #Input:
-#    <cleanDirectory> = directory of reference and sample fastq data (lambdaVirusInputs or agonaInputs)
-#    <workDirectory> = directory where the clean data is copied and results will be generated
-#    <compareDirectory> = directory containing correct result files (sample pileups, snplist, and snpma.fasta)
-#    <numFastq> = 1 or 2.  1=unpaired or 2=paired
-#    [platform] = HPC or can be omitted for non-HPC
+#    <cleanDirectory>   : directory of reference and sample fastq data (lambdaVirusInputs or agonaInputs)
+#    <workDirectory>    : directory where the clean data is copied and results will be generated
+#    <compareDirectory> : directory containing correct result files (sample pileups, snplist, and snpma.fasta)
+#    [platform]         : "torque" or can be omitted for non-HPC
 #
 #    This script expects the following directories and files:
 #    <cleanDirectory>/reference/<reference name>.fasta
@@ -31,67 +30,88 @@
 #    The script compares the generated results to the expected results ands emits differences.
 #Use example:
 #    cd test
-#    sh test_from_scratch.sh  ../snppipeline/data/lambdaVirusInputs  testLambdaVirus  ../snppipeline/data/lambdaVirusExpectedResults  2
-#    sh test_from_scratch.sh  ../snppipeline/data/agonaInputs  testAgona  ../snppipeline/data/agonaExpectedResults  1
+#    ./test_from_scratch.sh  ../snppipeline/data/lambdaVirusInputs  testLambdaVirus  ../snppipeline/data/lambdaVirusExpectedResults
+#    ./test_from_scratch.sh  ../snppipeline/data/agonaInputs  testAgona  ../snppipeline/data/agonaExpectedResults  torque
 #History:
 #   20140626-scd: Started.
 #   20140710-scd: Enhanced to conditionally support HPC (Torque).
+#   20140721-scd: Enhanced to support mix of paired and unpaired samples.
 #Notes:
 #Bugs:
 #
 
-if (($# < 4)) || (($# > 5)); then
-    echo usage: $0  cleanDirectory workDirectory compareDirectory numFastQ platform
+if (($# < 3)) || (($# > 4)); then
+    echo usage: $0  cleanDirectory workDirectory compareDirectory platform
     echo
-    echo 'cleanDirectory = directory of reference and sample fastq data (lambdaVirusInputs or agonaInputs)'
-    echo 'workDirectory = directory where the clean data is copied and results will be generated'
-    echo 'compareDirectory = directory containing correct result files (sample pileups, snplist, and snpma.fasta)'
-    echo 'numFastq = 1 or 2.  1=unpaired or 2=paired'
-    echo 'platform = HPC or can be omitted for non-HPC'
+    echo 'cleanDirectory   : directory of reference and sample fastq data (lambdaVirusInputs or agonaInputs)'
+    echo 'workDirectory    : directory where the clean data is copied and results will be generated'
+    echo 'compareDirectory : directory containing correct result files (sample pileups, snplist, and snpma.fasta)'
+    echo 'platform         : "torque" or can be omitted for regular non-HPC environments'
     exit 1
 fi
 
 CLEANDIR=$1
 WORKDIR=$2
 COMPAREDIR=$3
-NUMFASTQ=$4
-PLATFORM=$5
+PLATFORM=$4
 
-if (($# < 5)); then
+NUMCORES=$(grep -c ^processor /proc/cpuinfo)
+
+if (($# < 4)); then
     PLATFORM=Regular
 fi
 
 echo -e "\nStep 1 - Prep work"
-rm -rf $WORKDIR
-cp -r $CLEANDIR $WORKDIR
+mkdir -p $WORKDIR
+cp -v -u -r $CLEANDIR/* $WORKDIR
+find $WORKDIR -name "*.bt2" -type f -delete
+find $WORKDIR -name "*.fai" -type f -delete
+find $WORKDIR -name "*.pileup" -type f -delete
+find $WORKDIR -name "*.sam" -type f -delete
+find $WORKDIR -name "*.bam" -type f -delete
+find $WORKDIR -name "*.vcf" -type f -delete
+rm $WORKDIR/*
 cd $WORKDIR
 ls -d --color=never samples/* > sampleDirectoryNames.txt
-find samples -type f | grep fastq | sort -u > sampleFullPathNames.txt
+rm sampleFullPathNames.txt
+cat sampleDirectoryNames.txt | while read dir; do echo $dir/*.fastq >> sampleFullPathNames.txt; done
 sampleCount=$(cat sampleDirectoryNames.txt | wc -l)
 
 echo -e "\nStep 2 - Prep the reference"
 referencePath=$(ls reference/*.fasta)
 referenceBasePath=${referencePath%.fasta} # strip the file extension
-if [[ $PLATFORM == HPC ]]; then
+if [[ $PLATFORM == torque ]]; then
     prepReferenceJobId=$(echo prepReference.sh $referenceBasePath | qsub -d $WORKDIR -N job.prepReference -j oe)
 else
     prepReference.sh $referenceBasePath
 fi
 
-echo -e "\nStep 3 - Prep the samples"
-if [[ $PLATFORM == HPC ]]; then
-    cat sampleFullPathNames.txt | xargs -n $NUMFASTQ  > temp.prepSamples.input
-    prepSamplesJobId=$(echo | qsub -t 1-$sampleCount -d $WORKDIR -N job.prepSamples -j oe -W depend=afterok:$prepReferenceJobId << _EOF_
-    prepSamplesParameters=\$(cat temp.prepSamples.input | head -n \$PBS_ARRAYID | tail -n 1)
+echo -e "\nStep 3 - Align the samples to the reference"
+if [[ $PLATFORM == torque ]]; then
+    numAlignThreads=8
+    alignSamplesJobId=$(echo | qsub -t 1-$sampleCount -d $WORKDIR -N job.alignSamples -j oe -W depend=afterok:$prepReferenceJobId -l nodes=1:ppn=$numAlignThreads << _EOF_
+    alignSamplesParameters=\$(cat sampleFullPathNames.txt | head -n \$PBS_ARRAYID | tail -n 1)
+    alignSampleToReference.sh $numAlignThreads $referenceBasePath \$alignSamplesParameters
+_EOF_
+)
+else
+    cat sampleFullPathNames.txt | xargs --max-args=2 --max-lines=1 alignSampleToReference.sh $NUMCORES $referenceBasePath
+fi
+
+echo -e "\nStep 4 - Prep the samples"
+if [[ $PLATFORM == torque ]]; then
+    alignSamplesJobArray=${alignSamplesJobId%%.*}
+    prepSamplesJobId=$(echo | qsub -t 1-$sampleCount -d $WORKDIR -N job.prepSamples -j oe -W depend=afterokarray:$alignSamplesJobArray << _EOF_
+    prepSamplesParameters=\$(cat sampleDirectoryNames.txt | head -n \$PBS_ARRAYID | tail -n 1)
     prepSamples.sh $referenceBasePath \$prepSamplesParameters
 _EOF_
 )
 else
-    cat sampleFullPathNames.txt | xargs -n $NUMFASTQ prepSamples.sh $referenceBasePath
+    cat sampleDirectoryNames.txt | xargs -n 1 prepSamples.sh $referenceBasePath
 fi
         
-echo -e "\nStep 4 - Run snp pipeline (samtools pileup in parallel and combine alignment and pileup to generate snp matrix)"
-if [[ $PLATFORM == HPC ]]; then
+echo -e "\nStep 5 - Run snp pipeline (samtools pileup in parallel and combine alignment and pileup to generate snp matrix)"
+if [[ $PLATFORM == torque ]]; then
     cmd="create_snp_matrix.py -d ./ -f sampleDirectoryNames.txt -r $referencePath -l snplist.txt -a snpma.fasta -i True"
     prepSamplesJobArray=${prepSamplesJobId%%.*}
     createSnpMatrixJobId=$(echo $cmd | qsub -d $WORKDIR -N job.createSnpMatrix -j oe -W depend=afterokarray:$prepSamplesJobArray)
@@ -99,8 +119,8 @@ else
     create_snp_matrix.py -d ./ -f sampleDirectoryNames.txt -r $referencePath -l snplist.txt -a snpma.fasta -i True
 fi    
 
-echo -e "\nStep 5 - compare results"
-if [[ $PLATFORM == HPC ]]; then
+echo -e "\nStep 6 - compare results"
+if [[ $PLATFORM == torque ]]; then
     snpmaFileCount=0
     while ((snpmaFileCount == 0)); do
         qstat -a
@@ -109,7 +129,6 @@ if [[ $PLATFORM == HPC ]]; then
         snpmaFileCount=$(ls snpma.fasta 2>/dev/null | wc -l) # wait for the SNP matrix to exist
     done
 fi
-#if ! [[ $COMPAREDIR =~ ^[/~].*$ ]]; then
 if [[ $COMPAREDIR != /* ]] && [[ $COMPAREDIR != ~* ]]; then
     COMPAREDIR=../$COMPAREDIR # handle relative path
 fi